обновление с новой таблицей ПСЦ. Без фронта

+++ src/main/java/by/base/main/dao/TelegramChatQualityDAO.java
@@ -8,10 +8,10 @@ public interface TelegramChatQualityDAO {
 	
 	List<TelegramChatQuality> getChatIdList();
 	
-	boolean existsById(int chatId);
+	boolean existsById(Long chatId);
 	
 	void save(TelegramChatQuality telegramChatQuality);
 	
-	void deleteByChatId (int chatId);
+	void deleteByChatId (Long chatId);
 
 }
